refactor: alphabetise all the things and move styles to src folder
