+++ packages/author-profile/src/author-profile.js
@@ -1,27 +1,27 @@
 import React from "react";
+import get from "lodash.get";
 import { withPageState } from "@times-components/pagination";
 import {
   AuthorArticlesNoImagesProvider,
   AuthorArticlesWithImagesProvider
 } from "@times-components/provider";
 import { ratioTextToFloat } from "@times-components/utils";
-import get from "lodash.get";
-import AuthorProfileError from "./author-profile-error";
 import AuthorProfileContent from "./author-profile-content";
+import AuthorProfileError from "./author-profile-error";
 import { propTypes, defaultProps } from "./author-profile-prop-types";
 import authorProfileTrackingContext from "./author-profile-tracking-context";
 
 const AuthorProfile = ({
   author,
   error,
-  refetch,
   isLoading,
-  onArticlePress,
-  onTwitterLinkPress,
   page,
+  pageSize: initPageSize,
+  onArticlePress,
   onNext,
   onPrev,
-  pageSize: initPageSize,
+  onTwitterLinkPress,
+  refetch,
   slug
 }) => {
   if (error) {
@@ -31,13 +31,13 @@ const AuthorProfile = ({
   if (isLoading) {
     return (
       <AuthorProfileContent
-        isLoading
-        showImages
-        pageSize={initPageSize}
-        imageRatio={ratioTextToFloat("3:2")}
         articlesLoading
+        imageRatio={ratioTextToFloat("3:2")}
+        isLoading
         onTwitterLinkPress={() => {}}
+        pageSize={initPageSize}
         refetch={() => {}}
+        showImages
       />
     );
   }
@@ -74,25 +74,25 @@ const AuthorProfile = ({
         variables: { imageRatio = "3:2" }
       }) => (
         <AuthorProfileContent
-          isLoading={isLoading}
-          name={name}
+          articles={get(data, "articles.list", [])}
+          articlesLoading={articlesLoading}
           biography={biography}
-          uri={uri}
-          error={articlesError}
-          refetch={refetchArticles}
-          jobTitle={jobTitle}
-          twitter={twitter}
-          onTwitterLinkPress={onTwitterLinkPress}
           count={get(articles, "count", 0)}
+          error={articlesError}
+          imageRatio={ratioTextToFloat(imageRatio)}
+          isLoading={isLoading}
+          jobTitle={jobTitle}
+          name={name}
+          onArticlePress={onArticlePress}
           onNext={onNext}
           onPrev={onPrev}
+          onTwitterLinkPress={onTwitterLinkPress}
           page={page}
           pageSize={pageSize}
-          imageRatio={ratioTextToFloat(imageRatio)}
+          refetch={refetchArticles}
           showImages={hasLeadAssets}
-          articlesLoading={articlesLoading}
-          articles={get(data, "articles.list", [])}
-          onArticlePress={onArticlePress}
+          twitter={twitter}
+          uri={uri}
         />
       )}
     </SelectedProvider>
